First pass at adding round-trip validation
This tests monitors by actually sending them to datadog to pass their
black-box validation (which we fail fairly often). This creates a
monitor with a random prefix, then immediately deletes it if it was
created successfully.

This doesn't handle failures very well yet, and needs more test for
what happens when you test multiple new monitors.

Signed-off-by: Bradon Kanyid <bradon@kanyid.org>