Refactored and finished failure logic for round-trip validation
This now is useful for checking that our monitors are actually
acceptable by DataDog.

In a perfect world, we could have bonny do a
check job on this, but I don't think there's a good way to get bonny
the API secrets to do so. Instead, we could configure this as a
pre-commit hook to stop simple parsing errors from getting into the
monitors repo.

Signed-off-by: Bradon Kanyid <bradon@kanyid.org>